서블릿 11일차 : 서블릿 복습 - 게시글 수정 예외처리 적용

+++ servlet/servlet_project2/src/main/java/kr/kh/app/service/BoardServiceImp.java
@@ -152,23 +152,24 @@ public class BoardServiceImp implements BoardService {
 			return false;
 		}
 		
+		
+		if(nums != null) {
+			//삭제할 첨부파일 삭제
+			for(String numStr : nums) {
+				try {
+					int num = Integer.parseInt(numStr);
+					FileVO fileVo = boardDao.selectFile(num);
+					deleteFile(fileVo);
+				}catch (Exception e) {
+					e.printStackTrace();
+				}
+			}
+		}
+		
 		//첨부파일 추가
 		for(Part part : partList) {			
 			uploadFile(part, board.getBo_num());
 		}
-		
-		//첨부파일 삭제
-		for(String numStr : nums) {
-			try {
-				int num = Integer.parseInt(numStr);
-				FileVO fileVo = boardDao.selectFile(num);
-				deleteFile(fileVo);
-			}catch (Exception e) {
-				e.printStackTrace();
-			}
-		}
-		
-		
 		
 		//같으면 게시글 수정
 		return boardDao.updateBoard(board);
